requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client & Async File Operations  
requests==2.31.0
aiofiles==23.2.0
aiohttp==3.9.1

# Video and Audio Processing
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
openai-whisper==20231117

# Text Processing & NLP (for intelligent sentence segmentation)
regex==2023.10.3
jieba==0.42.1

# Image Processing
Pillow==10.1.0

# Configuration & Environment
python-dotenv==1.0.0

# Legacy Dependencies (可根据需要移除)
youtube-transcript-api==0.6.2

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1

# Development & Debugging
colorama==0.4.6